requirements.txt
pandas
pyarrow
kafka-python
redis
psycopg2-binary